[DateTimePicker] Move DateTimePickerView to the module where it's used
